rename: rename package names
전체 모듈에 대해서 기본 패키지 com.aireview.review로 통일
모든 패키지 기본 패키지 하위에 위치.
@ComponentScan으로 별도의 설정없이 모든 모듈의 bean scan가능
