+++ domain/src/test/java/com/aireview/review/SpringDomainTestApplication.java
@@ -1,0 +1,7 @@
+package com.aireview.review;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringDomainTestApplication {
+}
